adding DC/OS Master version
Adding this version is useful for testing and CI. It is absolutely not meant to be used for production. Your cluster is at risk to be automatically upgraded, failed installs and mismatch versions of agents and masters without a known rollback stragety. The master branch will always introduce changing flags, features and newer methods to deploy DC/OS. The maintainer of this repo is required to follow the changes required in master so a successful cluster will be deployed. When failed deployments occur, we fix the master version and redeploy. After an official DC/OS version is released, the version will be an official DC/OS number, then the changes in Master can be copied into that specific version number.